[CARBONDATA-3287]Remove the validation for same schema in a location and fix drop datamap issue
### Why this PR?
Currently we have a validation that if there are two carbondata files in a location with different schema, then we fail the query. I think there is no need to fail. If you see the parquet behavior also we cna understand.

Here i think failing is not good, we can read the latets schema from latest carbondata file in the given location and based on that read all the files and give query output. For the columns which are not present in some data files, it wil have null values for the new column.

But here basically we do not merge schema. we can maintain the same now also, only thing is can take latest schma.

### Points to Observe
1. one data file with columns a,b and c. 2nd file is with columns a,b,c,d,e. then can read and create table with 5 columns or 3 columns which ever is latest and create table(This will be when user does not specify schema). If he species table will be created with specified schema
2. Only **validation** happens is if the column name is same in both the data files present at location, and the datatype is different, then the query fails
3. When ffirst query is fired the dtamap is created for the table, and if new column is present in other data file, the the datamap is not updated as the table name will be same, so if column list is different we can drop datamap and create again

This closes #3121