[CARBONDATA-3298]Removed Log Message for Already Deleted Segments
Problem:
In old store, Create table and perform one Insert operation. Now,
update and delete that record, which marks that segment as "MARKED FOR DELETE".
Now run "clean files command" to delete the segment.
Note: In this case, Metadata folder doesn't contain segment file.
In new store, Refresh the table and again perform IUD operation. Now,
when "clean files" command is executed, We will check if physically segment file
exists or not, if not present, then we will log a warning message, as file not present.
If old store contains more segments, then for each segment, this log message will be
getting printed, which is not required.

Solution:
Removed log message for already deleted segments

This closes #3131