[CARBONDATA-3301]Fix inserting null values to Array<date> columns in carbon file format data load
Problem:
When carbon datasource table contains columns like complex column with Array or Array and data is inserted and queried, it gives null data for those columns.

Solution:
In file format case before the actual load, we hwt the internal row object from spark, and we convert the internal row to carbondata understandable object, so that obejvct for date will be of Integertype, So while inserting data only long case is handled and we were passing this interger value to parse in SimpleDateFormat, which throws exception and we were inserting null. SO handled for integer. In this case directly assign the surrogate key with this value.

This closes #3133