[CARBONDATA-3302] [Spark-Integration] code cleaning related to CarbonCreateTable command
What changes were proposed in this pull request?
Removed Extra check to validate whether the stream relation is not null , moreover condition can be optimized further,
currently the condition has path validation whether path is part of s3 file system and then system is checking
whether the stream relation is not null, this check can be added initially as this overall
condition has to be evaluated for stream table only if stream is not null.

This closes #3134