[CARBONDATA-3305] Added DDL to drop cache for a table
Added CarbonDropCacheCommand to drop all cache entries for a particular table.

usage: DROP METACACHE ON TABLE tableName
Dropping cache for child table is not supported, the table has to be parent table.
Running the above command will clear all the entries belonging to the table,
its index entries, its datamap entries and it's forward and reverse dictionary.

This closes #3138