[CARBONDATA-3307] Fix Performance Issue in No Sort
When creating the table without sort_columns and loading the data into it, it is generating more carbondata
files than expected. Now the no. of carbondata files is being generated based on the no. of threads launched.
Each thread is initialising its own writer and writing data.

Now we pass the same writer instance to all the threads, so all the threads will write the data to same file.

This closes #3140