[CARBONDATA-3313] count(*) is not invalidating the invalid segments cache
Problem:
If any segment is deleted the next query has to clear/invalidate the datamap cache for those invalid segments. But count(*) has not considered the invalid segments to clear the datamap cache.

Solution:
In count(*) flow, before clearing the datamap cache, check and add the invalid segments of that table.

This closes #3144