[CARBONDATA-3293] Prune datamaps improvement for count(*)
Problem:
(1) Currently for count (*) , the prune is same as select * query. Blocklet and ExtendedBlocklet are formed from the DataMapRow and that is of no need and it is a time consuming process.
(2) Checking the update/delete status all the time.

Solution:
(1) We have the blocklet row count in the DataMapRow itself, so it is just enough to read the count. With this count (*) query performance can be improved.
(2) No need to check the update/delete status all the time unless the table is not updated/deleted.

This closes #3148