[CARBONDATA-3322] [CARBONDATA-3323] Added check for invalid tables in ShowCacheCommand & Standard output on ShowCacheCommand on table
Problem 1:
After we alter table name from t1 to t2, SHOW METACACHE ON TABLE works for both old table name t1 and new table name t2.
Fix:
Added check for table.

Problem 2:
When SHOW METACACHE ON TABLE is executed and carbonLRUCAche is null, output is empty sequence, which is not standard.
Fix:
Return standard output even when carbonLRUCache is not initalised (null) with size for index and dictionary as 0.

This closes #3157