[CARBONDATA-3330] Fix Invalid Exception while clearing datamap from SDK carbon reader
problem:
When clearing the datamap in SDK carbon reader close(), it was always checking for schema file to build table. for SDK it will not find schema file and hence the exception.

cause : This was already an existing bug, now visible after #2878
Carbon table is required only for launching datamap clearing job, for SDK no need to launch the job

solution: Look for carbon table only if need to launch the datamap clearing job.

This closes #3162