[CARBONDATA-3331] Fix for external table in Show Metacache
Problem: In SHOW METACACHE command when an external table is queries upon, database size is more than ALL size for index column. External table is created upon a store on a table which is also present in current database.

Bug: Index size for database was being summed up blindly for all ths tables. So some cache entries were being summed up multiple times.

Solution: Compute the database index size while iterating over cache, when cache key contains path of one of the table's path.

This closes #3164