[CARBONDATA-3332] Blocked concurrent compaction and update/delete
Problem:
When update and compaction are executed concurrently then update is trying to update the contents of segment 0 whereas compaction has already marked the segment as COMPACTED. This compacted segment is of no use to the update command and therefore when trying to access the segment info from map it throws key not found.

Solution:
Block update and compaction concurrent operations.

This closes #3166