[CARBONDATA-3341] fixed invalid NULL result in filter query
Problem: When vector filter push down is true and the table contains a null value
then thegetNullBitSet method is giving an byte[]to represent null.
But there is no check for the value of the bitset.

Solution: Check if null bit set length is 0 then set the same to the chunkData.

This closes #3172