[CARBONDATA-3353 ]Fixed MinMax Based Pruning for Measure column in case of Legacy store
Why this PR needed?

Problem:
For table created and loaded with legacy store having a measure column, while building the page min max,
min is written as max and viceversa, so blocklet level minmax is wrong. With current version, when we query with filter on measure column, measure filter pruning is skipping some blocks and giving wrong results.

Solution:
Skip MinMax based pruning in case of legacy store for measure column.

This closes #3176