[CARBONDATA-3343] Compaction for Range Sort
Problem: To support Compaction for Range Sort in correct way as earlier it was grouping the ranges/partitions based on taskId which was not correct.

Solution: Combine all the data and create new ranges using Spark's RangePartitioner and using them give each
range to one task and apply the filter query to get the compacted segment.

This closes #3182