[CARBONDATA-3375] [CARBONDATA-3376] Fix GC Overhead limit exceeded issue and partition column as range column issue
Problem1 : When only single data item is present then it will be launched as one single task wich results in one executor getting overloaded.

Solution: When only a single range then we divide the splits and give to different tasks in order to ensure one executor does not overload.

Problem2 : When the range col is given as partitioned by column then compaction is failed because compaction goes to Range Column flow.

Solution: Added a check for Partition Table when range column is present so that it goes through the old flow and compaction passes.

This closes #3210