[CARBONDATA-3386] Concurrent Merge index and query is failing
Problem:
Concurrent merge index and query is failing. Load is triggered on a table, at the end of the load Merge index will be triggered. But this is triggered after the table status is updated as SUCCESS/PARTIAL SUCCESS for that segments. So for the concurrent query, this segment is available for query. Once the merge index is done, it deletes the index files, which are still referred by the query, this leads to the query failure.

Solution:
Trigger merge index before table status updation for all the operations like LOAD, COMPACT.

This closes #3221