[CARBONDATA-3309]Fix MV modular plan generation failure for spark 2.1
Problem:
In Spark2.1 version, we have Metastore relation instead of HiveTableRelation
as in current spark. So this case is not handled.

Solution
When it is MetastoreRelaton, get the catalogTable and get all the info from
it to form the ModularRelation, to form the modular plan. Reflection is used to fix this.

This closes #3257