[CARBONDATA-3445] Aggregate query empty list error fix.
Problem:
In Aggregate query, CountStarPlan throws head of empty list error.
When right node of join node has Aggregate node with the empty aggregate and group expression then it matches with the count() plan
and we try to get the Head of the aggregate expression assuming only one aggregate expression is present.
This happens only in Spark 2.3 so added an empty list check that if aggregate expression is empty, it will not treat as count() plan.

Aggregate query empty list error fix.

This closes #3300