[CARBONDATA-3452] Fix select query failure when substring on dictionary column with join
Problem: select query failure when substring on dictionary column with join.

Cause: when dictionary include is present, data type is updated to int from string in plan attribute.
so substring was unresolved on int column. Join operation try to reference this attribute which is unresolved.

Solution: skip updating datatype if dictionary is included in the plan

This closes #3306