[HOTFIX] Fixed date filter issue for fileformat
Problem:
In fileformat spark converts date to date object and sends to the carbon, carbon converts it to string and
then generates miliseconds. But there is a milli second gap of spark generated milli and carbon generated milli.
This causes date filters are not working properly.

Solution:
Convert the date to millis in spark side before give to carbon.

This closes #3327