[CARBONDATA-3473] Fix data size calcution of the last column in CarbonCli
When update last column chunk data size, current code use columnDataSize.add(fileSizeInBytes - footerSizeInBytes - previousChunkOffset) for every blocklet. This leads to wrong result for calculting the data size of the last column, especially when a carbon data file has multiple blocklet.

In this PR, we fix this problem and modify the calcultion by remarking the end offset of blocklet.

This closes #3330