[CARBONDATA-3482] Fixed NPE in Concurrent query
Problem: In case of concurrent queries if Q1 is loading cache and Q2 is removing from cache then Q2 may remove the segmentPropertiesIndex which Q1 has allocated and is about to access. This will cause NullPointerException .

Solution: Instead of adding index in BlockDataMap keep the reference of segmentPropertiesWrapper to be used.

This closes #3351