[CARBONDATA-3488] Check the file size after move local file to carbon path
Problem:
the row num saved in carbonindex file is non zero but the file size of relevant carbondata file is 0.

Solution:
In CarbonUtil.copyCarbonDataFileToCarbonStorePath, check the file size of carbon file whether is the same as the size fo local file after move local file to carbon path.

This closes #3350