[CARBONDATA-3452] dictionary include udf handle all the scenarios
Problem: select query failure when substring on dictionary column with join.
Cause: when dictionary include is present, data type is updated to int from
string in plan attribute. so substring was unresolved on int column.
Join operation try to reference this attribute which is unresolved.
Solution: Need to handle this for all the scenarios in CarbonLateDecodeRule

This closes #3358