[CARBONDATA-3494]Fix NullPointerException in drop table and Correct the document formatting
This closes #Problem:

Fix the formatting of the document in index server md file.
drop table is calling drop datamap command with force drop as true. Due to this table is removed from meta and physically. Then when processData is called for drop table, it tried to create carbonTable object by reading schema which causes NullPointerException.
Solution:

correct the formatting
Skip ProcessData if carbonTable is null

This closes #3359