[CARBONDATA-3496] Maintain the list of TableDataMap based on tableId instead of tableUniqueName
Problem:
TableNotFoundException is thrown from index server randomly for various file like schema & .carbondata.
Through code analysis it was found that if table cache is stale in the index executors and a query is fired then the cache can return a TableDataMap objetc which is old and the carbondata file would not exist for the same.
Solution:
Make the cache based on tableId instead of tableUniqueName to avoid stale cache access.

This closes #3374