[CARBONDATA-3502] Select query with UDF having Match expression inside IN expression Fails
Problem: Select query with UDF having Match expression inside
IN expression Fails with ArrayIndexOutOfBounds exception.

Cause: The expression should not be treated as Match expression,
instead should be treated as SparkUnknownExpression.

Solution: Removed the check for Match Expression as it was only added
for Lucene Search mode, which is no longer present.

This closes #3363