[CARBONDATA-3505] Drop database cascade fix
Problem: When 2 databases are created on same location and one of them is dropped
then the folder is also deleted from backend. If we try to drop the 2nd database
then it would try to lookup the other table, but the schema file would not exist
in the backend and the drop will fail.

Solution: Add a check to call CarbonDropDatabaseCommand only if the database
location exists in the backend.

This closes #3365