[CARBONDATA-3506]Fix alter table failures on parition table with hive.metastore.disallow.incompatible.col.type.changes as true
Problem:
In case of spark2.2 and above and , when we call alterExternalCatalogForTableWithUpdatedSchema to update the new schema to external catalog
in case of add column, spark gets the catalog table and then it itself adds the partition columns if the table is partition table for all the
new data schema sent by carbon, so there will be duplicate partition columns, so validation fails in hive
When the table has only two columns and one of them is partition column, then dropping non partition column is invalid because,
 if we allow it is like table with all columns as partition columns. So with the above property as true, drop column will fail to update the hive metastore.
in spark2.2 and above if the datatype change is done on partition column, with the above property as true, it also fails,
 as we are not sending partition column for schema alter in hive

Solution:
when sending the new schema to spark to update in catalog, do not send the partition columns in case of spark2.2 and above,
as spark will take care of adding parition columns to new schema sent by us.
In the above scenario of drop, do not allow drop column, if after dropping the specific column, if table has only partition columns.
Block the operation on datatype change on partition column on spark2.2 and above.

This closes #3367