[CARBONDATA-3507] Fix Create Table As Select Failure in Spark-2.3
Problem: Create table as select fails with Spark-2.3.
Cause: When creating the table location path the function
removes the "hdfs://" part from the path and then stores it,
due to which in later stages the file is treated as a Local Carbon File.
Solution: Get the original table path without removing the prefix.

This closes #3368