[CARBONDATA-3501] Fix update table with varchar column
Problem
Update on table with varchar column will throw exception

Analyse
In the loading part of update operation, it gets the isVarcharTypeMapping for each column in the order when table created. And this gives a hint for checking string length. It does not allow to exceeds 32000 char for a column which is not varchar type.

However when changing the plan for updating in CarbonIUDRule, it first deletes the old expression and appends the new one, which makes the order differ to table created. Such that the string length checking fail.

Solution
Keep the order as table created when modify update plan

This closes #3398