[CARBONDATA-3552]Fix dataload failure for column added in 1.1 which is a sort column in latest version
Problem:

When dimension is added in old version like carbon1.1, by default it will be sort column. In sort step we assume data will be coming as sort column in begining. But the added column is at last eventhogh sort column. So mapping for comparator and actual data mismatch happens as carbon row still contains the added column data at last. So data comparision fails and data load fails.

Solution:

While building the dataload configuration for loading data, rearrange the columns in orer to bring the sort column to beginning and non sort to last. So the carbonRow will be correct. But before the final merge setp, we update the sortRow based on schema, so again the sort column comes at end, so In final merger step, added one more comparator class, which will have the information about actual index of data in the Carbon row. Which helps to solve this issue.

This closes #3410