[CARBONDATA-3561] Fix incorrect results after execute delete/update operation if there are null values
Problem:
If there are null values in table, after execute delete/update sql, the results are incorrect when query with vector mode

Root cause:
In method 'fillVector' of 'LocalDictDimensionDataChunkStore', if the type of 'vector' is 'ColumnarVectorWrapperDirectWithDeleteDelta', the parameter 'counter' of 'ColumnarVectorWrapperDirectWithDeleteDelta' does not increase when the data is not null, so it will set the null value to the wrong index.

Solution:
Increase 'counter' of 'ColumnarVectorWrapperDirectWithDeleteDelta' when the data is not null.