[CARBONDATA-3579] Support merging index files when adding new partition
Normally, application will use Carbon SDK to write files into a
partition folder, then add the folder to partitioned carbon table.
If there are many threads writes to the same partition folder, there
will be many carbon index files, and it is not good for query
performance since all index files need to be read to spark driver.

So, a better way is to merge the index files when adding new partition
to carbon table.

This closes #3452