[CARBONDATA-3586][CARBONDATA-3587][CARBONDATA-3595]Adding valid segments into segments to be refreshed map before inserting segments to index server
Modification reason:
After select query the cache is doubled and the drop metacache is not removing the
cache in Spark 2.1 Carbon for the Index Server

Modification Content:
The preprimed segments were considered as invalid segments during the select query
as datamapstoremanager had no information of them and was sent to a different executor
during select query which was inturn doubling the cache. So the valid segments before
prepriming are sent into the map for them to be considered as valid segments during
select query and thus the query going into the same Index Server executor as during load.

This closes #3466