[CARBONDATA-3532] Support Query Rollup for MV TimeSeries Queries
Supported Query RollUp for MV TimeSeries queries.
How it is supported?
For each timeseries query, check if query can be rolled up from the existing datasets, by replacing the granularity in given user query. If query is rewritten and rolledUp, then add Select and Group by nodes on the rewritten query.

This closes #3495