[CARBONDATA-3602]Fix MV issues with session level operations
Problems:
1. when the same timeseries query is being executed from different
session, one session query is not being hit to datamap table.
When the query executed from different session, the semantic equals
of ScalaUDF expression fails, as object reference is different,
when the object equals is checked for that.

2. When same datamap on same table, but in some other database is
executed, fails with table not found error. This is because, when
first time the datamap catalog map is initialized, it contains the
session info, which wont be cleared when session get refreshed.

Solution:
1. make the TimeSeriesFunction as case class, as during run time
scala generates the equals method, which will check the other
equality when the object reference are not equal.

2. when the session info gets updated, clear the entry in map
and put again with new session info.

This closes #3497