[CARBONDATA-3623]: Fixed global sort compaction failure on timestamp column
Problem:
In Carbondata the timestamp is converted into long and stored in RDD.
While collect(action) is applied on dataframe it gives an error due to the confliction between data types that is - long in carbonScanRDD(from which dataframe is created) and Timestamp in schema.

Solution:
Called Dataset.ofRows(LogicalPlan) which returns the dataframe. Used Set command which sets all segments to be compacted. And Unset once the compaction is finished.

This closes #3515